requirements.txt
django==2.0.13
argon2-cffi==19.0.1
django-bootstrap4==0.0.8
django-misaka==2.1.1
django-braces==1.13.0
pillow==6.1.0
bcrypt==3.1.7